move customer calls to venue
