--- lib/customer_api/.env.example
@@ -1,2 +1,0 @@
-AUTH0_DOMAIN=""
-AUTH0_API_IDENTIFIER=""
