--- lib/customer_api/Dependencies
@@ -1,1 +1,0 @@
-!lib/protobuf
