--- lib/customer_api/Dockerfile
@@ -1,32 +1,0 @@
-# syntax = docker/dockerfile:experimental
-FROM golang:1.15-buster AS builder
-
-ENV GO111MODULE=on
-WORKDIR /src
-
-COPY .protobuf ./.protobuf
-
-COPY go.mod .
-COPY go.sum .
-
-RUN go mod download
-
-COPY . .
-
-RUN go test -tags=unit ./... \
-  && CGO_ENABLED=0 go build -a -o /main cmd/api/*.go
-
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-FROM alpine:3.12.0 as release
-
-RUN apk add --update \
-  curl \
-  tini \
-  ;
-
-COPY --from=builder /main /
-COPY --from=builder /src/internal/postgres/migrations ./internal/postgres/migrations
-
-EXPOSE 8888
-ENTRYPOINT ["/sbin/tini", "--"]
-CMD ["/main"]
