--- lib/customer_api/Dockerfile.dev
@@ -1,16 +1,0 @@
-FROM golang:1.15-buster
-
-ENV GO111MODULE=on
-WORKDIR /go/src
-
-COPY .protobuf ./.protobuf
-
-COPY go.mod .
-COPY go.sum .
-RUN go mod download
-
-RUN ["go", "get", "github.com/cespare/reflex"]
-
-COPY . .
-
-ENTRYPOINT ["reflex", "-c", "./reflex.conf"]
