--- lib/customer_api/docker-compose.yaml
@@ -1,39 +1,0 @@
-version: '3'
-services:
-  customer-api:
-    build:
-      context: .
-      dockerfile: Dockerfile.dev
-    ports:
-      - "8899:8899"
-    volumes:
-      - .:/go/src
-    environment:
-      - AUTH0_DOMAIN=${AUTH0_DOMAIN}
-      - AUTH0_API_IDENTIFIER=${AUTH0_API_IDENTIFIER}
-      - DATABASE_URL=postgres://user:password@customer-api-db/customer-api-db?sslmode=disable
-    depends_on:
-      - customer-api-db
-    networks:
-      main:
-        aliases:
-          - customer-api
-  customer-api-db:
-    image: "postgres"
-    container_name: "customer-api-db"
-    environment:
-      - POSTGRES_PASSWORD=password
-      - POSTGRES_DB=customer-api-db
-      - POSTGRES_USER=user
-    volumes:
-      - customer-api-db-data:/var/lib/postgresql/data
-    networks:
-      main:
-        aliases:
-          - customer-api-db
-volumes:
-  customer-api-db-data:
-networks:
-  main:
-
-#docker exec -it customer-api-db psql -U user -W customer-api-db customer-api-db
