--- lib/customer_api/reflex.conf
@@ -1,1 +1,0 @@
--r '(\.go$|go\.mod)' -s -- sh -c 'go test -tags=unit ./... && go build -a -o ./main cmd/api/main.go && ./main'
