--- lib/protobuf/src/customer/api/service.proto
@@ -1,40 +1,0 @@
-syntax = "proto3";
-
-package customer.api;
-
-option go_package = "github.com/cobbinma/booking-platform/lib/protobuf/autogen/lang/go/customer/api";
-
-service CustomerAPI {
-  rpc IsAdmin(IsAdminRequest) returns (IsAdminResponse);
-  rpc AddAdmin(AddAdminRequest) returns (AddAdminResponse);
-  rpc RemoveAdmin(RemoveAdminRequest) returns (RemoveAdminResponse);
-}
-
-message IsAdminRequest {
-  string venueId = 1;
-  string email = 2;
-}
-
-message IsAdminResponse {
-  bool isAdmin = 1;
-}
-
-message AddAdminRequest {
-  string venueId = 1;
-  string email = 2;
-}
-
-message AddAdminResponse {
-  string venueId = 1;
-  string email = 2;
-}
-
-message RemoveAdminRequest {
-  string venueId = 1;
-  string email = 2;
-}
-
-message RemoveAdminResponse {
-  string email = 1;
-}
-
