--- lib/venue_api/internal/postgres/.snapshots/Test_Repository-add_table_successfully
@@ -1,1 +1,0 @@
-(*models.Table)(id:"b31a9f99-3f64-4ee9-af27-45b2acd36d86" name:"test table" capacity:4)
