--- lib/venue_api/internal/postgres/.snapshots/Test_Repository-get_venue_successfully
@@ -1,1 +1,0 @@
-(*models.Venue)(id:"b31a9f99-3f64-4ee9-af27-45b2acd36d86" name:"Test Venue" openingHours:{dayOfWeek:1 opens:"10:00" closes:"20:00"} openingHours:{dayOfWeek:2 opens:"10:00" closes:"22:00"})
